infra/runpod/requirements-ltx.txt
# RunPod LTX-Video 13B worker dependencies

# RunPod SDK
runpod==1.7.1

# Core ML frameworks
torch==2.4.0
torchvision==0.19.0

# Diffusers (LTX-Video pipeline)
diffusers==0.31.0
transformers==4.45.0
accelerate==1.0.0

# HuggingFace Hub (모델 선택적 다운로드용)
huggingface_hub>=0.25.0

# Image / video processing
Pillow==10.4.0
imageio[ffmpeg]==2.35.1
numpy==1.26.4

# HTTP client for downloading keyframe images
requests==2.32.3

# Sentencepiece (T5 tokenizer 필요)
sentencepiece==0.2.0
protobuf==5.28.0
